[ refactor ] hashtag구현부 리팩토링
